dev-requirements.txt
django==2.0.9
channels==2.1.7
bleach==3.1.0
channels-redis==2.3.3
psycopg2==2.7.7
Sphinx==1.8.4
sphinx-rtd-theme==0.4.3
twine==1.13.0
django-tenants==2.2.3
coverage==4.5.3
daphne==2.2.5
pytest-cov==2.7.1
pytest-asyncio==0.10.0
pytest-django==3.4.8
selenium==3.141.0
